+++ src/backend/joanie/core/templatetags/extra_tags.py
@@ -2,6 +2,7 @@
 
 from django import template
 from django.contrib.staticfiles import finders
+from django.utils.translation import gettext as _
 
 from joanie.core.utils import image_to_base64
 
@@ -15,3 +16,15 @@ def base64_static(path):
     if full_path:
         return image_to_base64(full_path, True)
     return ""
+
+
+@register.filter
+def join_and(items: list):
+    """A template tag filter to join a list of items in human-readable way."""
+    comma_join_threshold = 2
+    if len(items) > comma_join_threshold:
+        return _("{:s} and {:s}").format(
+            ", ".join(map(str, items[:-1])), str(items[-1])
+        )
+
+    return _(" and ").join(map(str, items))