+++ src/backend/joanie/tests/core/test_templatetags_extra_tags.py
@@ -3,7 +3,7 @@ Test suite for extra template tags of the joanie core app
 """
 from django.test import TestCase
 
-from joanie.core.templatetags.extra_tags import base64_static
+from joanie.core.templatetags.extra_tags import base64_static, join_and
 
 
 class TemplateTagsExtraTagsTestCase(TestCase):
@@ -28,3 +28,23 @@ class TemplateTagsExtraTagsTestCase(TestCase):
         """
         static_filepath = "unknown-static-file.txt"
         self.assertEqual(base64_static(static_filepath), "")
+
+    def test_templatetags_extra_tags_join_and(self):
+        """
+        The template tags `join_and` should join a list of items in a human-readable way
+        """
+        # - Test with a single item
+        items = ["Joanie"]
+        self.assertEqual(join_and(items), "Joanie")
+
+        # - Test with two items
+        items += ["Richie"]
+        self.assertEqual(join_and(items), "Joanie and Richie")
+
+        # - Test with three items
+        items += ["Fonzie"]
+        self.assertEqual(join_and(items), "Joanie, Richie and Fonzie")
+
+        # - Test with four items
+        items += ["Marsha"]
+        self.assertEqual(join_and(items), "Joanie, Richie, Fonzie and Marsha")